Implemented speculative contacts collision detection
It worked briefly but then I started adding features..
(adding the type of collision required to move the dude on the ground
i.e. it detected the collisions in the right place but wasn't passing
enough information around so that we could usefully handle the collision
later on - we need to decuple handling collisions from detecting them).

I'm tired now so I'm packing it in it's not working properly at the moment though.