Fixed the "falling through the floor bug"
When an obj would nearly have hit the floor and simultaneously did hit
another particle object the first particle would be marked as a potential
collider - but not a potential collider with the floor.  OK so far.
The first particle now gets moved step by step because it's a potential collider.

The problem occurs because of an out by one error when moving objects step
by step.  Since the algo mistakenly takes an extra step at
the end it can move the first particle into a position where it would
now be colliding with the floor but doesn't because there's no potential
collision with the floor to detect.

A real bastard to find.